backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlmodel==0.0.19
aiosqlite==0.20.0

# AI / LangChain
langchain==0.2.0
langchain-google-genai==1.0.5
langchain-community==0.2.0
google-generativeai==0.5.4
openai==1.30.0

# Vector DB & embeddings
faiss-cpu==1.9.0.post1
langchain-text-splitters==0.2.0

# PDF parsing
pymupdf==1.25.5

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
